backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "mao-backend"
version = "0.1.0"
description = "MAO 营销多智能体协同编排平台 - 后端服务"
requires-python = ">=3.11"
dependencies = [
    # Web 框架
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "python-multipart>=0.0.9",
    # 数据库
    "sqlalchemy>=2.0.30",
    "aiomysql>=0.2.0",
    "alembic>=1.13.1",
    "pymysql>=1.1.0",
    # 缓存
    "redis[hiredis]>=5.0.4",
    # 消息队列
    "aiokafka>=0.10.0",
    # 配置管理
    "pydantic>=2.7.1",
    "pydantic-settings>=2.3.0",
    "python-dotenv>=1.0.1",
    # 大模型
    "openai>=1.30.1",
    "tiktoken>=0.7.0",
    # 安全
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    # HTTP 客户端
    "httpx>=0.27.0",
    # 工具
    "ulid-py>=1.1.0",
    "croniter>=2.0.5",
    "jinja2>=3.1.4",
    "structlog>=24.2.0",
    "tenacity>=8.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.7",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
    "ruff>=0.4.7",
    "mypy>=1.10.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
strict = true
